refactor: 원시값 포장 객체를 비교 가능하게 변경

+++ src/main/java/racingcar/model/dto/CarDto.java
@@ -12,7 +12,7 @@ public class CarDto {
     }
 
     public static CarDto from(Car car) {
-        return new CarDto(car.name(), car.distance());
+        return new CarDto(car.name(), car.distance().intValue());
     }
 
     public String name() {
